Tidy up, check platform and add comments in qlocale_win.cpp
MS says we should stop using LCID, so make a note about that.
This file should only be compiled when Q_OS_WIN is defined, so
don't bother with #if-ery to check that.
Remove two redundant forward-declarations.

Record what qt_localeFromLCID is for - undocumented private API
exported for the use of a platform plugin - since that isn't
immediately obvious to someone only looking at QLocale.

Change-Id: I81ad945bb92d4f81a3ca2aaf97a945635b137a2e
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>