wasm: make EventCallback use addEventListener()
EventCallback would previously set the on<Event> property
on the event target, which is a singleton property where
there can be only one event handler.

This was OK if the event target was owned by Qt, for
example the canvas element, where we could guarantee
that there was only one event handler.

However this approach fell through when registering
event handlers for global event targets, such as for
window.onLanguageChange, where setting a singleton event
handler may conflict with other users.

Fix this by using the addEventListener() API instead,
which has a variant which takes an event listener object,
which gives us the ability to provide C++ context for
the listener. 

The C++ context in this case is a std::function that
contains the event handler callback. Attempts to pass
this type to JavaScript was met with some resistance
from Emscripten, so as a late night hack pass it as
a uintptr_t for now.

Change-Id: I1a547b49af467882ae4f57f8d909ffdff0be6b51

Reviewed-by: Piotr Wierciński <piotr.wiercinski@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>