Add QtPrivate::ArrowProxy
... use it in QKeyValueIterator.

This is somewhat simpler than the old QKeyValueIterator::pointer.

I don't think the pointer::operator*() could have ever worked, because
it returns value_type&, but is const.

That leaves the defaulted copy and move SMFs as a difference (ArrowProxy
is using aggregate initialization, so doesn't need any), and the fact
that operator->() is const.

Change-Id: I80b9c5f696de6ae30f3939166b205b9213eac57e
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>