wasm: make preload_qml_imports take a source path
This was running qmlimportscanner on ".", which in some
cases would recurse down the qt/ symlink in the app
build directory (used with dynamic linking on wasm),
and find all of the imports from the Qt installation.

Make it take a path instead. Users can then provide
a path to the QML sources which will be passed to
qmlimportscanner

Change-Id: Ib5175e5dc1d26875c42f5a3e286314b7d602c9fe

Reviewed-by: Piotr Wierciński <piotr.wiercinski@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>