windows: Avoid infinite recursion with certain fonts
Amends 922d195020d54d7e599d135f6a5e0338100e08f1.

Another infinite recursion was found in the case of populating
font families for cases where the typographical name is different
from the populated name.

The specific font triggering this had two "preferred name" entries
in its name data. This was one for Chinese (PRC) and another for
Chinese (Taiwan). Our GDI backend does not do matching based on
locale, but will prefer the English if available, otherwise it
will just pick the first one. The font in question does not have
the English preferred name.

For this particular font, we would select the Chinese (Taiwan)
name as preferred and since it had not been populated, we would
populate this. However, on Chinese (PRC) locale, Windows would
report this according to the Chinese (PRC) name. We would once
again translate this to Chinese (Taiwan) and go into an infinite
recursion populating it.

The quick fix is to mark the preferred family as populated before
entering the recursion, so that we do not re-populate it a second
time later.

Ideally, the font database would match the preferred name based
on locale, but since we are moving away from the GDI font database
and this is already handled better with the DirectWrite database,
just fixing the recursion is sufficient here.

[ChangeLog][Windows] Fixed an issue where an infinite recursion
could occur if the system had a font with multiple preferred names
in non-English languages.

Pick-to: 6.6 6.5 6.2 5.15
Fixes: QTBUG-118238
Change-Id: I6ded369f1c908e51c0ba2ad9127538faf07e192e
Reviewed-by: Lars Knoll <lars@knoll.priv.no>
Reviewed-by: Liang Qi <liang.qi@qt.io>
(cherry picked from commit 11344f57235b5c8149d231c1331f86bcff51bd50)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>